+++ Framework/ConvivaAppAnalytics.xcframework/tvos-arm64/ConvivaAppAnalytics.framework/Headers/CATNetworkReqTrackingConfiguration.h
@@ -17,6 +17,8 @@ NS_ASSUME_NONNULL_BEGIN
 
 @property(assign, readonly) int32_t rbl; //max supported request/response body length in KB
 
+@property(assign, readonly) int32_t maxQueryLength; //max supported request query params length in KB
+
 @property(strong, readonly) NSArray *blocklist;
 
 @property(strong, readonly) NSArray *collectattr;